feat: Enhance inward and purchase order functionalities with new models and associations, update GRN and GRNItem fields, and improve data handling in controllers
- create New Sales order create and fetch API
- update Inward controller
- update PO controller
- modify GRNItems model
- update POItem model

+++ src/models/grnItems.model.js
@@ -15,22 +15,22 @@ function GRNItem(sequelize) {
             allowNull: false
         },
         ordered_qty: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull: false,
             defaultValue: 0
         },
         shortage_qty: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull: false,
             defaultValue: 0
         },
         damage_qty: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull: false,
             defaultValue: 0
         },
         received_qty: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(10, 2),
             allowNull: false,
             defaultValue: 0
         },
